refactor: API 엔드포인트 모듈별 분리 및 라우트 구조 개선
- portfolio, category, asset, asset-history 모듈로 분리
    - asset, asset-history는 추후 분리 예정
- RESTful API 라우트 구조 적용 (복수형 통일)
- 계층적 URL 구조로 일관성 확보
